webimar-api/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
crispy-bootstrap4==2024.10
Django==5.2
django-cors-headers==4.7.0
django-crispy-forms==2.4
djangorestframework==3.16.0
idna==3.10
lxml==5.4.0
numpy==2.2.5
psycopg2-binary==2.9.10
pykml==0.2.0
python-decouple==3.8
requests==2.32.3
shapely==2.1.0
sqlparse==0.5.3
urllib3==2.4.0
